# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CargoLink Backend
  description: Smart Freight Matching & Tracking System — backend API. This is the live-generated source
    of truth for the API contract; the committed, versioned export lives at shared/openapi/openapi.yaml
    and is checked for drift by backend/tests/test_openapi_contract.py.
  version: 0.1.0
paths:
...
  /shippers/profile:
    post:
      tags:
      - profiles
      summary: Create/complete the calling shipper's profile
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipperProfileResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipperProfileCreateRequest'
    get:
      tags:
      - profiles
      summary: Read the calling shipper's own profile
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipperProfileResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
  /transporters/profile:
    post:
      tags:
      - profiles
      summary: Create/complete the calling transporter's profile
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransporterProfileResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransporterProfileCreateRequest'
    get:
      tags:
      - profiles
      summary: Read the calling transporter's own profile
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransporterProfileResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
  /vehicles:
    post:
      tags:
      - vehicles
      summary: Register a vehicle owned by the calling transporter
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleCreateRequest'
    get:
      tags:
      - vehicles
      summary: List the calling transporter's own vehicles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleListResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
  /vehicles/{vehicle_id}:
    get:
      tags:
      - vehicles
      summary: Read a single vehicle
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: vehicle_id
        in: path
        required: true
        schema: &id002
          type: string
          format: uuid
        description: Unique vehicle identifier.
  /loads:
    post:
      tags:
      - loads
      summary: Post a new cargo load
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadCreateResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadCreateRequest'
...
  /loads/{load_id}/matches:
    get:
      tags:
      - matching
      summary: Get candidate vehicle matches for a load
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadMatchesResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: load_id
        in: path
        required: true
        schema: *id002
        description: The load to find matches for.
      - name: radius_km
        in: query
        required: false
        schema:
          type: number
          exclusiveMinimum: 0
          maximum: 500
          default: 50
        description: Search radius around the load's pickup point, in km.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: Maximum number of candidates to return.
...
  /notifications:
    get:
      tags:
      - notifications
      summary: List the calling user's notifications
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
  /notifications/{notification_id}/read:
    patch:
      tags:
      - notifications
      summary: Mark a notification as read
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '400':
          description: Validation error.
          content:
            application/json:
              schema: *id001
        '401':
          description: Missing or invalid authentication.
          content:
            application/json:
              schema: *id001
        '403':
          description: Authenticated but not authorized for this action.
          content:
            application/json:
              schema: *id001
        '404':
          description: Resource not found.
          content:
            application/json:
              schema: *id001
        '409':
          description: Conflict with current resource state.
          content:
            application/json:
              schema: *id001
        '429':
          description: Rate limited.
          content:
            application/json:
              schema: *id001
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema: *id001
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      parameters:
      - name: notification_id
        in: path
        required: true
        schema: *id002
        description: Notification to mark read.
...
components:
  schemas:
...
    ValidationError:
      type: object
      properties:
        loc:
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
        msg:
          type: string
        type:
          type: string
      required:
      - loc
      - msg
      - type
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ShipperProfileCreateRequest:
      type: object
      properties:
        company_name:
          type: string
          minLength: 1
          maxLength: 200
          description: Registered or trading company name.
        gstin:
          type:
          - string
          - 'null'
          pattern: ^[0-9]{2}[A-Z]{5}[0-9]{4}[A-Z]{1}[1-9A-Z]{1}Z[0-9A-Z]{1}$
          description: 15-character GST Identification Number. Optional — not every shipper is GST-registered.
          examples:
          - 22AAAAA0000A1Z5
      required:
      - company_name
    ShipperProfileResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: The shipper's user id.
        company_name:
          type: string
          description: Registered or trading company name.
        gstin:
          type:
          - string
          - 'null'
          description: 15-character GST Identification Number, if provided.
      required:
      - user_id
      - company_name
    TransporterProfileCreateRequest:
      type: object
      properties:
        license_no:
          type: string
          minLength: 1
          maxLength: 50
          description: Driving/operator license number.
      required:
      - license_no
    TransporterProfileResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: The transporter's user id.
        license_no:
          type: string
          description: Driving/operator license number.
        rating_avg:
          type: number
          minimum: 0
          maximum: 5
          default: 0.0
          description: Rolling average rating (0-5), fed back from the ratings module (Cluster H.2).
      required:
      - user_id
      - license_no
    VehicleType:
      type: string
      enum:
      - mini_truck
      - container_truck
      - trailer
      - other
    GeoPoint:
      type: object
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
          description: Latitude in decimal degrees (WGS84).
        lng:
          type: number
          minimum: -180
          maximum: 180
          description: Longitude in decimal degrees (WGS84).
      required:
      - lat
      - lng
    VehicleCreateRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/VehicleType'
        capacity_kg:
          type: number
          exclusiveMinimum: 0
          description: Maximum cargo weight capacity, in kilograms.
        capacity_volume:
          type: number
          exclusiveMinimum: 0
          description: Maximum cargo volume capacity, in cubic meters.
        route_pref:
          type:
          - string
          - 'null'
          maxLength: 200
          description: Free-text preferred route/operating corridor, e.g. 'Mumbai-Pune'.
      required:
      - type
      - capacity_kg
      - capacity_volume
    VehicleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique vehicle identifier.
        transporter_id:
          type: string
          format: uuid
          description: Owning transporter's user id.
        type:
          $ref: '#/components/schemas/VehicleType'
        capacity_kg:
          type: number
          description: Maximum cargo weight capacity, in kilograms.
        capacity_volume:
          type: number
          description: Maximum cargo volume capacity, in cubic meters.
        route_pref:
          type:
          - string
          - 'null'
          description: Free-text preferred route/operating corridor.
        current_location:
          anyOf:
          - $ref: '#/components/schemas/GeoPoint'
          - type: 'null'
          description: Last known GPS location, if the vehicle has reported one.
        last_ping_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Timestamp of the last location ping received.
      required:
      - id
      - transporter_id
      - type
      - capacity_kg
      - capacity_volume
    VehicleListResponse:
      type: object
      properties:
        vehicles:
          type: array
          items:
            $ref: '#/components/schemas/VehicleResponse'
          description: The calling transporter's registered vehicles.
      required:
      - vehicles
    CargoType:
      type: string
      enum:
      - general
      - fragile
      - hazardous
      - refrigerated
    LoadStatus:
      type: string
      enum:
      - open
      - matched
      - accepted
      - in_transit
      - delivered
      - cancelled
    LoadCreateRequest:
      type: object
      properties:
        weight:
          type: number
          exclusiveMinimum: 0
          description: Cargo weight in kilograms.
        cargo_type:
          $ref: '#/components/schemas/CargoType'
        source:
          $ref: '#/components/schemas/GeoPoint'
        destination:
          $ref: '#/components/schemas/GeoPoint'
        deadline:
          type: string
          format: date-time
          description: Latest acceptable pickup deadline.
      required:
      - weight
      - cargo_type
      - source
      - destination
      - deadline
    LoadCreateResponse:
      type: object
      properties:
        load_id:
          type: string
          format: uuid
          description: Unique identifier for the newly posted load.
        status:
          $ref: '#/components/schemas/LoadStatus'
      required:
      - load_id
      - status
...
    MatchCandidate:
      type: object
      properties:
        vehicle_id:
          type: string
          format: uuid
          description: Candidate vehicle's id.
        distance_km:
          type: number
          minimum: 0
          description: Road distance from vehicle to pickup point, in km (OSRM; falls back to Haversine
            if OSRM is degraded — Cluster C.1).
        capacity_fit:
          type: boolean
          description: Whether the vehicle's capacity satisfies the load's weight/volume.
        eta:
          type: integer
          minimum: 0
          description: Estimated time to pickup, in seconds.
        score:
          type: number
          description: Composite match score (higher is better); see Clusters C.2/H.2 for scoring inputs.
      required:
      - vehicle_id
      - distance_km
      - capacity_fit
      - eta
      - score
    LoadMatchesResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchCandidate'
          description: Candidate vehicles for this load, sorted by score descending. Empty list (not an
            error) when nothing qualifies.
      required:
      - matches
...
    NotificationType:
      type: string
      enum:
      - booking_confirmation
      - pickup_confirmation
      - delay_alert
      - delivery_confirmation
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique notification identifier.
        type:
          $ref: '#/components/schemas/NotificationType'
        message:
          type: string
          description: Human-readable notification text.
        sent_at:
          type: string
          format: date-time
          description: When this notification was created/sent.
        read:
          type: boolean
          default: false
          description: Whether the recipient has marked this notification read.
      required:
      - id
      - type
      - message
      - sent_at
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
          description: The calling user's notifications, most recent first.
      required:
      - notifications
...